requirements.txt
# MK Restaurants Parser & Agent Requirements

# Data processing
pandas>=2.0.0
xlrd>=2.0.0
openpyxl>=3.1.0

# GUI/Tray App
pystray>=0.19.0
Pillow>=10.0.0

# File watching
watchdog>=3.0.0

# Supabase client
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
